LIBSEARCH-81 inital commit for AS searcher
This is the first pass at the umd search Quick Search ArchivesSpace
Seacher. For searching.

This takes a request, passes it to the AS PUI, scrapes the results, and
displays them in the searchumd app.

https://issues.umd.edu/browse/LIBSEARCH-81

LIBSEARCH-81 Add .ruby files and some stylistic changes

This adds the .ruby-* files and makes some stylistics tweaks.

https://issues.umd.edu/browse/LIBSEARCH-81

LIBSEARCH-81 Make en.yml use more generic ArchivesSpace term

This replaces "The Archives" with the more generic and accurate
ArchivesSpace term.

https://issues.umd.edu/browse/LIBSEARCH-81